R177 ①b: a zoom is a similarity, not a convergence — and one ruler, shared
Two follow-ups to the tilt fix, both found by running the earlier rounds' own browser specs.

1. #R175's dolly came back. gEye scaled the LOOK DISTANCE by k but left the target's elevation
   alone, which is "dolly toward a fixed target" — and after any tilt that target is in the sky, so
   the camera converges ON it. Reproduced exactly the number #R175 recorded for the broken build:
   z12 Tokyo, pitch 110, wheel-zooming 2.3 levels, target frozen at 14,099 m, eye 8,373 -> 12,955 m,
   i.e. climbing while zooming in. #R175 settled that a zoom is a SIMILARITY about the map point
   under the centre: the eye AND the target both scale by k. Measured after: 1673.1943819879984
   against an expected 1673.1943819879998.

   (The wheel-zoom case reaches the anchored branch, not the travel branch — with the pointer at the
   canvas centre MapLibre's zoom-around-the-pointer does not move the centre at all.)

2. The ruler now lives in ONE file (tests/helpers/camera-ruler.js) and both r176.spec.js and
   r177.spec.js install it, for the same reason gEye/gSolve is one function: r176.spec.js had its
   own copy of #R176's correction, which is how four rounds of drift went unreported.

A/B'd against main to attribute the regression honestly: r175.spec.js passes on 18bafd4 and failed
on this branch before this commit.

27 passing across r175/r176/r177.
